Hide publish data from menu
